pyproject.toml
[project]
name = "welcometogermany"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.6",
    "fastapi>=0.95",
    "httpx>=0.24",
    "pydantic>=2.0",
    "pydantic-settings>=1.1",
    "pytest>=7.0",
    "python-dotenv>=1.0",
    "python-jose[cryptography]>=3.3.0",
    "requests>=2.28",
    "uvicorn[standard]>=0.22",
]
